requirements.txt
anthropic>=0.96.0
python-dotenv
rich
click
bandit
safety
pygithub
